feat(support-ticket-backend): refactor validateExtraFields to be a wrapper function
